requirements.txt
Django>2.1
djangorestframework
django-auth-ldap
uwsgi
